implemented add playlist and listviews
We can now see our playlists, songs in a selected playlist (currently can not be chosen), and songs in our library.

if we click new under playlists you can type in a name, and it will create a new file under the playlists directory to be used similar to "music".